Tinkering with themes and config.
Just seeing what code works and what doesn't. This should by no means be
considered final. I am going to need to see what the smart way to handle
this would be. The best way would probably be to break down config and
theme(s) into separate files. But that is going to take some more
dedicated time of thinking.