bug fix: Interpreter now runs at proper speed.
Prior to this the interpreter was running incredibly slow and I had no
idea why. I had completely neglected to add a proper timing system in so
the timers were getting dropped to 0 almost immediately.
Now the interpreter runs properly at the proper speed with the timers
decrementing as they should.
The rate at which the instructions are ran are also configurable with a
new line line added in config.toml and the corresponding field in it's
.rs file.